requirements.txt
torch
torchvision
numpy
pillow
opencv-python
matplotlib
scikit-learn
tqdm
flask
flask_cors
